requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
jinja2==3.1.4
httpx==0.28.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
aiofiles==24.1.0
babel==2.16.0
